Move HTTPApplication to SilverStripe\Control namespace
