+++ src/Core/Config/ConfigLoader.php
@@ -26,7 +26,7 @@ class ConfigLoader
      */
     public static function inst()
     {
-        return self::$instance ? self::$instance : self::$instance = new self();
+        return self::$instance ? self::$instance : self::$instance = new static();
     }
 
     /**
@@ -92,7 +92,7 @@ class ConfigLoader
         $manifest = clone $this->getManifest();
 
         // Create new blank loader with new stack (top level nesting)
-        $newLoader = new self;
+        $newLoader = new static;
         $newLoader->pushManifest($manifest);
 
         // Activate new loader
